replace old info card with new one
