+++ src/routes/routes.js
@@ -12,7 +12,7 @@ import Signup from "../pages/Signup";
 import PageNotFound from "../pages/PageNotFound";
 // can remove this after testing IconsDataStructure is done
 // import Testing from "../pages/Testing";
-import infoCard_test from "../components/infoCard_test";
+import InfoCard from "../components/infoCard";
 
 const theme = createMuiTheme({
   palette: {
@@ -35,7 +35,7 @@ function Routes() {
             <Route path="/admin" component={ifAuth(AuthLandingPage)} />
             <Route path="/login" component={Login} />
             <Route path="/signup" component={Signup} />
-            <Route path="/testing" component={infoCard_test} />
+            <Route path="/testing" component={InfoCard} />
             <Route component={PageNotFound} />
           </Switch>
         </ThemeProvider>
